add checks to make sure multiple snappyzone instances aren't spawned, only kill active instances

+++ src/snappyzones/process.py
@@ -2,13 +2,40 @@ import os
 import signal
 import sys
 
+from functools import lru_cache
+
 from .service import Service
 
 
 HERE = os.path.abspath(os.path.dirname(__file__))
+PID_FILE = os.path.join(HERE, '.pid')
+
+
+def _file_pid():
+    if os.path.isfile(PID_FILE):
+        with open(PID_FILE, 'r') as f:
+            pid = int(f.read())
+        return pid
+    return None
+
+
+def _check_pid(pid): 
+    if not pid:
+        return False       
+    try:
+        os.kill(pid, 0)
+    except OSError:
+        return False
+    return True
 
 
 def launch_background_process(*args, **kwargs):
+    # if there is a process already running, kill it
+    _pid = _file_pid()
+    if _check_pid(_pid):
+        print("Background Snappy Zones process found: killing process...")
+        stop_background_process(_pid)
+        
     print("STARTING...")
     try:
         pid = os.fork()
@@ -19,7 +46,7 @@ def launch_background_process(*args, **kwargs):
         sys.exit(1)
 
     pid = str(os.getpid())
-    
+
     with open(os.path.join(HERE, '.pid'), 'w') as f:
         f.write(pid)
 
@@ -28,10 +55,17 @@ def launch_background_process(*args, **kwargs):
     service.listen()
     
 
-def stop_background_process(*args, **kwargs):
-    with open(os.path.join(HERE, '.pid'), 'r') as f:
-        pid = int(f.read())
-    print(f"STOPPING...")
-    os.kill(pid, signal.SIGTERM)
-    print(f"STOPPED PID: {pid}")
+def stop_background_process(pid=None):
+    if not pid:
+        pid = _file_pid()
+    if _check_pid(pid):
+        print(f"STOPPING...")
+        os.kill(pid, signal.SIGTERM)
+        print(f"STOPPED PID: {pid}")
+        os.remove(PID_FILE)
+    else:
+        print(
+            "No background process was found. "
+            "no action will be taken."
+        )
